* Fixed some bugs in the primality tests
* Made some more things threadsafe

* Made changes to multimodular multiplication such as using files and to thetaproduct example

git-svn-id: https://flint.svn.sourceforge.net/svnroot/flint/trunk@1781 2531bd8e-7522-0410-9704-d59f193d18de

+++ QS/tinyQS.c
@@ -175,7 +175,7 @@ unsigned long tiny_collect_relations(linalg_t * la_inf, QS_t * qs_inf, poly_t * 
 
 int F_mpz_factor_tinyQS(F_mpz_factor_t * factors, mpz_t N)
 {
-    int small_factor, i;
+    ulong small_factor, i;
 
 
     small_factor = F_mpz_factor_tinyQS_silent(factors, N);
